[FEAT] Picker: 코드 파일 분리 및 그룹화 #13
